resolver: move resolve_workspace into its own module
lib.rs is now 277 lines — just the `Resolver` struct, `ResolveTask`,
module wiring, and re-exports. The 1800-line BFS implementation lives
in resolve.rs as its own `impl Resolver` block.

Also extracts the direct-deps seeding loop into a pure
`seed_direct_deps` helper (55 lines out of the main method body).

Progress across this branch:

  lib.rs: 6504 -> 2545 -> 2091 -> 277 lines

Further structural improvement of resolve_workspace itself is blocked
on refactoring the macro-captured shared state (queue, visited,
resolved, in_flight, etc.) into an explicit BfsState struct — worth
its own PR.
